+++ src/clear_store.ts
@@ -1,4 +1,4 @@
-import { _wrap_a1 } from '@ctx-core/function'
+import { wrap_a_ } from '@ctx-core/function'
 import { each } from '@ctx-core/array'
 import type { Writable } from './writable'
 /**
@@ -7,7 +7,7 @@ import type { Writable } from './writable'
 export function clear_store<Val extends unknown = unknown>(
 	stores:Writable<Val>|Writable<Val>[], val:Val
 ):void {
-	const store_a1 = _wrap_a1<Writable<Val>>(stores)
+	const store_a1 = wrap_a_<Writable<Val>>(stores)
 	each(store_a1, (store:Writable<Val>)=>store.set(val))
 }
 export { clear_store as clear__store }