+++ src/readable_set_ctx_.ts
@@ -3,7 +3,7 @@ import { Readable, readable } from './readable'
 import { get } from './get'
 import type { StartStopNotifier } from './StartStopNotifier'
 import type { Subscriber } from './Subscriber'
-export function _readable_set_ctx<Val extends unknown = unknown>(
+export function readable_set_ctx_<Val extends unknown = unknown>(
 	initial:Val, readable_fn = readable
 ):readable_set_ctx_T<Val> {
 	let set = undefined as maybe_undefined<_readable_set_ctx_set_T<Val>>
@@ -20,4 +20,7 @@ export type _readable_set_ctx_set_T<Val extends unknown = unknown> = Subscriber<
 export interface readable_set_ctx_T<Val extends unknown = unknown> {
 	store:Readable<Val>
 	set:_readable_set_ctx_set_T<Val>
+}
+export {
+	readable_set_ctx_ as _readable_set_ctx,
 }