+++ src/readable_stores_type.ts
@@ -1,2 +1,0 @@
-import type { Readable } from './readable'
-export type readable_stores_type<Val extends unknown = unknown> = Readable<Val>|Readable<Val>[]