+++ src/readable_stores_val_T.ts
@@ -1,6 +1,6 @@
-import type { readable_stores_type } from './readable_stores_type'
+import type { readable_stores_T } from './readable_stores_T'
 import type { Readable } from './readable'
-export type readable_stores_val_type<Store extends readable_stores_type<unknown>> =
+export type readable_stores_val_T<Store extends readable_stores_T<unknown>> =
 	Store extends Readable<infer N0_Val>[]
 	? N0_Val
 	: Store extends Readable<infer N1_Val>