+++ src/set_store_.ts
@@ -4,7 +4,7 @@ import type { Writable } from './writable'
  * Returns a function to set the given store using the value returned by `setter`.
  * This is useful in conjunction with [subscribe](#subscribe).
  */
-export function _set_store<Val extends unknown = unknown>(
+export function set_store_<Val extends unknown = unknown>(
 	store:Writable<Val>,
 	setter = (v:Val)=>v
 ):set_store_T<Val> {
@@ -15,4 +15,7 @@ export function _set_store<Val extends unknown = unknown>(
 			: setter)
 }
 export type set_store_T<Val extends unknown = unknown> = (val:Val)=>void
-export { _set_store as _set__store }
+export {
+	set_store_ as _set_store,
+	set_store_ as _set__store, 
+}