+++ src/stores_infer_T.ts
@@ -1,3 +1,3 @@
 import type { Stores } from './Stores'
-export type stores_infer_type<Store extends Stores<unknown>> =
+export type stores_infer_T<Store extends Stores<unknown>> =
 	Store extends (infer Val) ? Val : never