+++ src/writable$_C.ts
@@ -8,10 +8,8 @@ export class writable$_C<Val extends unknown = unknown> extends readable$_C<Val>
 	}
 	readonly set:Writable_set_T<Val> = this.store.set
 	readonly update:Writable_update_T<Val> = this.store.update
-	public get $():Val {
-		return get(this.store)
-	}
-	public set $(val) {
-		this.store.set(val)
-	}
+	public get $():Val {return get(this.store)}
+	public set $(val) {this.store.set(val)}
+	public get _():Val {return get(this.store)}
+	public set _(val) {this.store.set(val)}
 }