+++ src/writable_stores_T.ts
@@ -1,0 +1,2 @@
+import type { Writable } from './writable'
+export type writable_stores_T<Val extends unknown = unknown> = Writable<Val>|Writable<Val>[]