+++ src/writable_stores_type.ts
@@ -1,2 +1,0 @@
-import type { Writable } from './writable'
-export type writable_stores_type<Val extends unknown = unknown> = Writable<Val>|Writable<Val>[]